% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aois.R
\name{add_aois}
\alias{add_aois}
\title{Map the x and y positions of looks to Areas of Interest.}
\usage{
add_aois(x, aois, default_onscreen = "tracked")
}
\arguments{
\item{x}{a dataframe of looking data}

\item{aois}{an AOI or list of AOIs}

\item{default_onscreen}{default label to use for a look onscreen that does
not fall into an AOI. Default is \code{"tracked"}}
}
\value{
an updated dataframe
}
\description{
Map the x and y positions of looks to Areas of Interest.
}
\details{
The function current assumes the conventions used in our lab. It will create
a column called \code{GazeByAOI} with the label of the AOI for each look. It does
this by checking whether the columns \code{XMean} and \code{YMean} fall into to the
boundaries of the AOI.

Offscreen looks receive \code{NA}.
}
