% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{adjust_times_around_zero}
\alias{adjust_times_around_zero}
\title{Adjust time values around 0}
\usage{
adjust_times_around_zero(data, time_col = "Time", fps = 60, ties = "first")
}
\arguments{
\item{data}{a dataframe of eyetracking data for a single trial or a grouped
dataframe where the groups define a single trial.}

\item{time_col}{name (string) of the column with time value. Defaults to
\code{"Time"}.}

\item{fps}{the eyetracking sampling rate. Defaults to 60 frames per second.}

\item{ties}{how to break ties when the smallest times are equally close to
zero. Default is \code{"first"} so that the tie \code{c(-1, 1)} is aligned to
\code{c(0, 2)}.}
}
\value{
the dataframe with updated time values
}
\description{
Adjust time values around 0
}
