% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_looks}
\alias{aggregate_looks}
\title{Aggregate looks}
\usage{
aggregate_looks(data, resp_def, formula)
}
\arguments{
\item{data}{a long data frame of looking data}

\item{resp_def}{a "response definition" list that describes eyetracking
response values. The list should have elements named "primary", "others",
"elsewhere", and "missing".}

\item{formula}{an aggregation formula. The lefthand terms will be grouping
variables, and the righthand term is the column with eyetracking responses.}
}
\value{
a dataframe of the grouping columns with columns plus columns with
the number of responses and the proportion/se of looks to the primary
response type and the proportion of missing data
}
\description{
This function uses an aggregation formula.
}
\seealso{
\code{aggregate_looks2} for a version that does not use a formula.
}
