% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_looks2}
\alias{aggregate_looks2}
\title{Alternative function for aggregating looks}
\usage{
aggregate_looks2(data, resp_def, ..., resp_var)
}
\arguments{
\item{data}{a long data frame of looking data}

\item{resp_def}{a "response definition" list that describes eyetracking
response values. The list should have elements named "primary", "others",
"elsewhere", and "missing".}

\item{...}{Grouping columns.}

\item{resp_var}{Name of the column that contains eyetracking responses.}
}
\value{
a dataframe of the grouping columns with columns plus columns with
the number of responses and the proportion/se of looks to the primary
response type and the proportion of missing data
}
\description{
Alternative function for aggregating looks
}
