% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_looks_}
\alias{aggregate_looks_}
\title{Standard-evaluation version of `aggregate_looks`}
\usage{
aggregate_looks_(data, resp_def, grouping, resp_col)
}
\arguments{
\item{data}{a long data frame of looking data}

\item{resp_def}{a "response definition" list that describes eyetracking
response values. The list should have elements named "primary", "others",
"elsewhere", and "missing".}

\item{grouping}{Grouping columns.}

\item{resp_col}{Name of the column that contains eyetracking responses.}
}
\description{
Standard-evaluation version of `aggregate_looks`
}

