% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{assign_bins}
\alias{assign_bins}
\title{Assign bins for downsampling looking data}
\usage{
assign_bins(data, bin_width = 3, time_var, ..., bin_col = ".bin",
  na_location = "tail", partial = FALSE)
}
\arguments{
\item{data}{a dataframe of looking data}

\item{bin_width}{the number of items to put in each bin. Default is 3.}

\item{time_var}{the name of the column representing time}

\item{...}{grouping variables}

\item{bin_col}{name of the column to add. Defaults to \code{".bin"}.}

\item{na_location}{Where to assign \code{NA} bin numbers. \code{"head"} and \code{"tail"}
respectively put the \code{NA} elements at the head and tail of the vector;
\code{"split"} alternates between \code{"tail"} and \code{"head"}.}

\item{partial}{whether to exclude values that don't fit evenly into bins.
Defaults to \code{FALSE}, so that the user is warned if a bin is incomplete.}
}
\value{
the original dataframe with an added column of bin numbers. The
dataframe will be sorted by the grouping and time variables.
}
\description{
Assign bins for downsampling looking data
}
