% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawiz.R
\name{convert_datawiz_code_to_aoi}
\alias{convert_datawiz_code_to_aoi}
\title{Convert to DataWiz codes to AOI names}
\usage{
convert_datawiz_code_to_aoi(xs)
}
\arguments{
\item{xs}{a vector of DataWiz codes (-, ., 0, 1)}
}
\value{
the vector with NA for "-", "Target" for "1", "Distractor" for "0", and
  "tracked" for ".".
}
\description{
Convert to DataWiz codes to AOI names
}

