% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{cycle_response_def}
\alias{cycle_response_def}
\title{Create complementary response definitions}
\usage{
cycle_response_def(response_def)
}
\arguments{
\item{response_def}{a response definition to use a template for other
definitions.}
}
\value{
a list of response definitions where each member of
\code{c(response_def$primary, response_def$others)} is used as the primary
response.
}
\description{
In the typical response definition, there is a primary response compared to
other competitors. Oftentimes, we are interested in also comparing each of
the competitors to the other images. This function quickly assembles a full
cycle of response definitions.
}
\examples{
# Create one definition
def <- create_response_def(
  primary = 1,
  others = c(5, 8, 9),
  elsewhere = 0,
  missing = NA
)

# Create the full cycle of response definitions
cycle_response_def(def)
}
