% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littlelisteners-package.r
\name{example_files}
\alias{example_files}
\title{Locate the path of example eyetracking files}
\usage{
example_files(which = 1)
}
\arguments{
\item{which}{index of the batch of example files to load}
}
\value{
Paths to a banch of examples files bundled with the
\code{littlelisteners} package.
}
\description{
Locate the path of example eyetracking files
}
\details{
This function is a wrapper over \code{\link[=system.file]{system.file()}}  to locate the
paths to bundled eyetracking data. These files are used to test or demonstrate
functionality of the package.

The following sets of files are included:
\enumerate{
\item \code{Coartic_Block1_001P00XS1} - Data for a block of trials from an eyetracking
performed with a Tobii Eyetracker in an Eprime experiment.
\item \code{Coartic_Block2_001P00XS1} - Data for a second block of trials from the
above experiment.
}
}
