% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{find_frequent_interval}
\alias{find_frequent_interval}
\title{Find the most frequent interval of values}
\usage{
find_frequent_interval(xs, min_freq = 0.8)
}
\arguments{
\item{xs}{some set of values (times)}

\item{min_freq}{the minimum frequency of times to keep}
}
\value{
a list with the minimum and maximum values that have a frequency of
at least 80\%.
}
\description{
When we have trials of varying lengths, it can be helpful to find the
interval of time shared by most of the trials.
}
\examples{
times <- c(
  -5:10, -10:10, -8:10, -8:10, -8:11, -8:13, -8:4,
  -5:12, -10:9
)
find_frequent_interval(times)
}
\keyword{internal}
