% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littlelisteners-package.r
\docType{data}
\name{four_image_data}
\alias{four_image_data}
\title{Example data from a Visual World experiment}
\format{
A data frame with 20,910 rows and 25 variables
}
\usage{
four_image_data
}
\description{
Example data from a Visual World experiment
}
\keyword{datasets}
