% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littlelisteners-package.r
\docType{package}
\name{littlelisteners}
\alias{littlelisteners-package}
\alias{littlelisteners}
\title{littlelisteners.}
\description{
This package houses frequently used functions for working with hand-coded eyetracking data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.tjmahr.com/littlelisteners/}
}

}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID})

}
\keyword{internal}
