% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawiz.R
\name{melt_datawiz}
\alias{melt_datawiz}
\title{Convert DataWiz data into long format}
\usage{
melt_datawiz(df, key_col = "Time", value_col = "Look")
}
\arguments{
\item{df}{a dataframe created by reading a datawiz file}

\item{key_col}{the name of the new column that holds the time values}

\item{value_col}{the name of the new column that holds the looking data at
each time sample}
}
\value{
a long data-frame
}
\description{
DataWiz files have several columns F0, F33, F67, etc. for each time sample.
This function converts a dataframe from such a file into a long format, where
there is a single time column and single column of gaze responses.
}
