% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazedata.R
\name{read_gazedata}
\alias{read_gazedata}
\title{Load a \code{.gazedata} file for an experiment}
\usage{
read_gazedata(gazedata_path, eyes = "both", means_need_both = FALSE,
  apply_corrections = TRUE)
}
\arguments{
\item{gazedata_path}{path to the \code{.gazedata} file that is to be parsed.}

\item{eyes}{string describing which eye(s) should be selected for the
\code{Mean} columns. Valid options are \code{"both"}, \code{"left"}, and
\code{"right"}. Defaults to \code{"both"}. If \code{"left"} is selected,
then only the left eye is used to calculate the \code{XMean}, \code{YMean},
etc. columns.}

\item{means_need_both}{logical value indicating if both eyes are required to
compute \code{Mean} columns. Defaults to \code{FALSE}. If \code{FALSE},
\code{NA} values are ignored, so for example, \code{XMean} could be
computed from an \code{XLeft} of .25 and an \code{XRight} of \code{NA}.}

\item{apply_corrections}{whether to do low-level adjustments like coding
offscreen looks as NA, negative pupil diameters to NA, negative distances
to NA, and flip y-axis so the origin is the lower-left corner. Defaults to
\code{TRUE}. Only used as \code{FALSE} in case "raw" data is needed.}
}
\value{
A dataframe containing the parsed gazedata. Each row of the dataframe
contains the eye-tracking data for a single frame of time recorded during
the experiment.
}
\description{
Loads \code{.gazedata} file created by an Eprime experiment running on a
Tobii eyetracker, and performs typical data reduction on that file.
}
\details{
We extract the columns the following columns: \code{TrialId}, \code{RTTime},
\code{XGazePosLeftEye}, \code{XGazePosRightEye}, \code{YGazePosLeftEye},
\code{YGazePosRightEye}, \code{DistanceLeftEye}, \code{DistanceRightEye},
\code{DiameterPupilLeftEye} and \code{DiameterPupilRightEye}.

Once these column values are loaded, we make three modifications to the
gazedata (when \code{apply_corrections} is TRUE).

\enumerate{
\item Gaze measurements with \code{Validity} codes greater than or equal to
1 are replaced with NA values.

\item X,Y gaze values are defined in screen proportions. Values that fall
outside \link{0,1} are outside of the boundaries of the screen and therefore are
nonsensical. Replace them with \code{NA}. We perform a similar correction
on pupil diameters and eye-distances by replacing negative
values with \code{NA}.

\item The origin of the screen is the upper-left-hand corner of the screen.
Flip the y-values so that the origin is in a more familiar position in the
lower-left-hand corner of the screen. This way, low y values are closer to
the bottom of the screen.

\item Compute the mean x, y, distance and diameter values for the left and
right eyes. \code{NA} values are ignored when computing the mean, so the
pair \code{(XLeft = NA, XRight = .5)} yields \code{XMean = .5}.
}
}
\references{
\href{http://bit.ly/1AtKyhR}{Tobii Toolbox for Matlab: Product
Description & User Guide}
}
