% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{create_response_def}
\alias{create_response_def}
\title{Create a response definition}
\usage{
create_response_def(primary, others, elsewhere = NULL, missing = NA,
  id = NULL)
}
\arguments{
\item{primary}{the primary response of interest}

\item{others}{other responses of interest}

\item{elsewhere}{responses to ignore}

\item{missing}{responses that indicate missing data. Defaults to NA.}
}
\value{
a list
}
\description{
Create a response definition
}
