% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolate_looks}
\alias{interpolate_looks}
\title{AOI-based gaze interpolation}
\usage{
interpolate_looks(x, window, fps, response_col, interp_col, fillable,
  missing_looks)
}
\arguments{
\item{x}{a dataframe of grouped eyetracking data. Each row should be a single
frame of eyetracking. Use \code{dplyr::group_by()} to set the grouping
columns for the data. The groups should specify a single trial of
eyetracking data.}

\item{window}{maximum amount of missing data (milliseconds) that can be
interpolated. Only spans of missing data with less than or equal to this
duration will be interpolated}

\item{fps}{number of eyetracking frames (dataframe rows) per second}

\item{response_col}{name of the column with the eyetracking response data}

\item{interp_col}{name of a column to add to the dataframe. This column
records whether each frame was interpolated (TRUE) or not (FALSE)}

\item{fillable}{values in the response column where interpolation is legal.
These would typically be AOI locations.}

\item{missing_looks}{values that can be imputed}
}
\description{
Fills in windows of missing data if the same AOI is fixated at
beginning and end of the missing data window. For example, the sequence
\code{"Target", NA, NA, "Target"} would be interpolated to be \code{"Target", "Target", "Target", "Target"}.
}
\details{
Use \code{window} to constrain the duration of missing data windows that
can be filled. We conventionally use 150ms because we would not expect
someone to shift their gaze from Image A to Image B to Image A in that
amount of time.
}
