% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{create_response_def}
\alias{create_response_def}
\title{Create a response definition}
\usage{
create_response_def(primary, others, elsewhere = NULL, missing = NA,
  label = NULL)
}
\arguments{
\item{primary}{the primary response of interest}

\item{others}{other responses of interest}

\item{elsewhere}{responses to ignore}

\item{missing}{responses that indicate missing data. Defaults to \code{NA}.}

\item{label}{optional label for the response definition. Defaults to the
value of \code{primary}.}
}
\value{
a \code{response_def} object
}
\description{
A response definition controls how \code{aggregate_looks()} works.
}
\details{
To deal with eyetracking data in a generic way, we need a way to describe
eyetracking responses. We assume that there are four basic gaze types.
\itemize{
\item Primary responses: A gaze to a primary or target image.
\item Other responses: Gazes to competing images.
\item Elsewhere looks: A gaze that is onscreen but not a primary or
other response. Typically, this occurs when the participant is
shifting between images.
\item Missing looks: A missing or offscreen gaze.
}

A \emph{response definition} is a programmatic way of describing these response
types, and it allows \code{aggregate_looks()} to map gaze data onto looking
counts.
}
\examples{
create_response_def(
  label = "looks to target",
  primary = "Target",
  others = c("PhonologicalFoil", "SemanticFoil", "Unrelated"),
  elsewhere = "tracked",
  missing = NA)
}
